requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.2
django-filter==2.2.0
django-heroku==0.3.1
django-oauth-plus==2.2.9
django-oauth2-provider==0.2.6.1
django-rest-framework==0.1.0
django-tastypie==0.14.3
djangorestframework==3.11.0
fcm-django==0.3.4
Flask==1.0.2
get==2019.4.13
gunicorn==20.0.4
httpie==2.0.0
httplib2==0.16.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
MarkupSafe==1.1.1
oauth2==1.9.0.post1
post==2019.4.13
psycopg2==2.8.5
psycopg2-binary==2.8.2
public==2019.4.13
pycparser==2.19
pyfcm==1.4.7
Pygments==2.5.2
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
shortuuid==0.5.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.7
views==0.3
virtualenv==16.7.9
Werkzeug==0.15.2
whitenoise==5.0.1
APScheduler~=3.6.3
PyMySQL~=0.9.3